+++ CPPScripts/Input/InputManagerWindows.cpp
@@ -39,7 +39,7 @@ namespace ZXEngine
 		mMouseX = xPos;
 		mMouseY = yPos;
 #endif
-		EventManager::GetInstance()->FireEvent((int)EventType::UPDATE_MOUSE_POS, to_string(mMouseX) + "|" + to_string(mMouseY));
+		EventManager::GetInstance()->FireEvent(EventType::UPDATE_MOUSE_POS, to_string(mMouseX) + "|" + to_string(mMouseY));
 	}
 
 	void InputManagerWindows::UpdateMouseScroll(float xOffset, float yOffset)
@@ -48,7 +48,7 @@ namespace ZXEngine
 		if (EditorInputManager::GetInstance()->CheckCurMousePos() != EditorAreaType::Game)
 			return;
 #endif
-		EventManager::GetInstance()->FireEvent((int)EventType::UPDATE_MOUSE_SCROLL, to_string(yOffset));
+		EventManager::GetInstance()->FireEvent(EventType::UPDATE_MOUSE_SCROLL, to_string(yOffset));
 	}
 
 	void InputManagerWindows::UpdateKeyInput()
@@ -90,11 +90,11 @@ namespace ZXEngine
 		SHORT state = GetAsyncKeyState(id);
 
 		if ((state & 0x8000) && mButtonState[(int)button] == 1)
-			EventManager::GetInstance()->FireEvent((int)e, ""); // Press
+			EventManager::GetInstance()->FireEvent((uint32_t)e, ""); // Press
 		else if ((state & 0x8000) && mButtonState[(int)button] == 0)
-			EventManager::GetInstance()->FireEvent((int)e + 1, ""); // Down
+			EventManager::GetInstance()->FireEvent((uint32_t)e + 1, ""); // Down
 		else if (!(state & 0x8000) && mButtonState[(int)button] == 1)
-			EventManager::GetInstance()->FireEvent((int)e + 2, ""); // Up
+			EventManager::GetInstance()->FireEvent((uint32_t)e + 2, ""); // Up
 
 		mButtonState[(int)button] = (state & 0x8000) ? 1 : 0;
 	}
@@ -106,11 +106,11 @@ namespace ZXEngine
 		string pos = to_string(mMouseX) + "|" + to_string(mMouseY);
 
 		if ((state & 0x8000) && mButtonState[(int)button] == 1)
-			EventManager::GetInstance()->FireEvent((int)e, pos); // Press
+			EventManager::GetInstance()->FireEvent((uint32_t)e, pos); // Press
 		else if ((state & 0x8000) && mButtonState[(int)button] == 0)
-			EventManager::GetInstance()->FireEvent((int)e + 1, pos); // Down
+			EventManager::GetInstance()->FireEvent((uint32_t)e + 1, pos); // Down
 		else if (!(state & 0x8000) && mButtonState[(int)button] == 1)
-			EventManager::GetInstance()->FireEvent((int)e + 2, pos); // Up
+			EventManager::GetInstance()->FireEvent((uint32_t)e + 2, pos); // Up
 
 		mButtonState[(int)button] = (state & 0x8000) ? 1 : 0;
 	}
